% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fpsim.R
\name{add_fpsim}
\alias{add_fpsim}
\title{Add dyadic foreign policy similarity measures to your data}
\usage{
add_fpsim(data, keep)
}
\arguments{
\item{data}{a data frame with appropriate \pkg{peacesciencer} attributes}

\item{keep}{an optional parameter, specified as a character vector, about what
dyadic foreign policy similarity measure(s) the user wants returned from this function. If
\code{keep} is not specified, the function returns all 14 dyadic foreign policy similiarity measures
calculated by Haege (2011). Otherwise, the function subsets the underlying data to just what the
user wants and merges in that.}
}
\value{
\code{add_fpsim()} takes a (dyad-year, leader-dyad-year) data frame and
adds information about the dyadic foreign policy similarity, based on several measures calculated and offered by Frank Haege.
}
\description{
\code{add_fpsim()} allows you to add a variety of dyadic foreign policy similarity measures to your
(dyad-year, leader-dyad-year) data frame
}
\details{
For the dyad-year (and leader-dyad-year) data, there must be some kind of information loss in order to reduce the disk space
data like these command. In this case, all calculations are rounded to three decimal spots. I do not think this to be terribly
problematic, though I admit I do not like it. If this is a problem for your research question (though I can't imagine it would be),
you may want to consider not using this function for dyad-year or leader-dyad-year data.

Be mindful that the data are fundamentally dyad-year and that extensions to leader-level data should be understood
as approximations for leaders-dyads in a given dyad-year.

The data this function uses are directed dyad-year and the merge is a left-join, making this function agnostic about
whether your dyad-year (or leader-dyad-year) data are directed or non-directed.
}
\examples{
\dontrun{
# just call `library(tidyverse)` at the top of the your script
library(magrittr)
# The function below works, but depends on running `download_extdata()` beforehand.
cow_ddy \%>\% add_fpsim()

# Select just kappa and pi from binary alliance data
cow_ddy \%>\% add_fpsim(keep=c("kappaba", "piba"))

}
}
\references{
\subsection{The Main Source of the Data}{

For any use of these data whatsoever, please cite Haege (2011). Data are version 2.0.
\itemize{
\item Haege, Frank M. 2011. "Choice or Circumstance? Adjusting Measures of Foreign Policy Similarity for Chance Agreement."
\emph{Political Analysis} 19(3): 287-305.
}
}

\subsection{Citations for the Particular Similarity Measure You Choose}{

Additional citations depend on what particular measure of similarity you're using,
either Signorino and Ritter's (1999) "S", Cohen's (1960) kappa and Scott's (1955) pi. Haege (2011) is
part of a chorus arguing against the use of "S", though "S" measures are included in these data if you elect
to ignore the chorus and use this mesure. Haege (2011) instead argues for a chance-corrected measure of dyadic
foreign policy similarity, either Cohen's (1960) kappa or Scott's (1955) pi.
\itemize{
\item Cohen, Jacob. 1960. "A Coefficient of Agreement for Nominal Scales." \emph{Educational and Psychological Measurement} 20(1): 37-46.
\item Scott, William A. 1955. "Reliability of Content Analysis: The Case of Nominal Scale Coding." \emph{Public Opinion Quarterly} 19(3): 321--5.
\item Signorino, Curtis S. and Jeffrey M. Ritter. "Tau-b or Not Tau-B: Measuring the Similarity of Foreign Policy Positions." 43(1): 115--44.
}
}

\subsection{Citations for the Underlying Data Informing the Similarity Measure}{

Haege (2011) also suggests you cite the underlying data informing the similarity measure, whether it is UN voting or alliances. In his case,
he recommended a Voeten citation from 2013 and the alliance data proper. In the case of the alliances, I know Gibler's (2009)
book is recommended even if the alliance data have since been updated (and reflected in this measure). In the UN voting data, my
understanding is the 2017 paper in \emph{Journal of Conflict Resolution} is also the preferred citation.
\itemize{
\item Bailey, Michael A., Anton Strezhnev, and Erik Voeten. 2017. "Estimating the Dynamic State Preferences from United Nations Voting Data."
\emph{Journal of Conflict Resolution} 61(2): 430--456.
\item Gibler, Douglas M. 2009. \emph{International Military Alliances, 1648-2008}. Washington DC: CQ Press.
}
}
}
\author{
Steven V. Miller
}
