% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-leaders.R
\docType{data}
\name{archigossums}
\alias{archigossums}
\title{Archigos Yearly Leader Turnover: A Summary}
\format{
A data frame with 14707 observations on the following 7 variables.
\describe{
\item{\code{gwcode}}{a numeric vector for the Gleditsch-Ward state code}
\item{\code{year}}{a numeric vector for a referent year}
\item{\code{leadertransition}}{a dummy variable indicating a leader transition in a given year}
\item{\code{irregular}}{a dummy variable indicating an irregular leader transition in a given year}
\item{\code{n_leaders}}{an integer for the number of leaders in a given year}
\item{\code{jan1obsid}}{a character vector for the observation ID of the head of state on Jan. 1 of the referent year}
\item{\code{dec31obsid}}{a character vector for the observation ID of the head of state on Dec. 31 of the referent year}
}
}
\usage{
archigossums
}
\description{
These are yearly summaries of leader turnover from the Archigos data, for use
in \code{add_archigos()}
}
\details{
Consult \code{archigos} in the same data frame for more information about the data.
}
\references{
Goemans, Henk E., Kristian Skrede Gleditsch, and Giacomo Chiozza. 2009. "Introducing
Archigos: A Dataset of Political Leaders" \emph{Journal of Peace Research}
46(2): 269--83.
}
\keyword{datasets}
