% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{citations}
\alias{citations}
\title{Citations for Data/Functions Used in \code{peacesciencer}}
\format{
A data frame with two variables:
\describe{
\item{\code{data_function}}{the data or function used or referenced in the package}
\item{\code{citation}}{an appropriate (text) citation you should include in your manuscript}
}
}
\usage{
citations
}
\description{
This is a master list of references for data/functions used in this package. Do check it out to make sure you're faithfully citing what you're using.
}
\keyword{datasets}
