% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cow_ddy}
\alias{cow_ddy}
\title{A directed dyad-year data frame of Correlates of War state system members}
\format{
A data frame with the following 3 variables.
\describe{
\item{\code{ccode1}}{a numeric vector for the Correlates of War state code for the first state}
\item{\code{ccode2}}{a numeric vector for the Correlates of War state code for the second state}
\item{\code{year}}{a numeric vector for the year}
}
}
\usage{
cow_ddy
}
\description{
This is a complete directed dyad-year data frame of Correlates of War
state system members. I offer it here as a shortcut for various other functions when
I am working on new additions and don't want to invest time in waiting for
\code{\link[=create_dyadyears]{create_dyadyears()}} to run. As a general rule, this data frame is
updated after every calendar year to include the most recently concluded
calendar year.
}
\details{
Data are a quick generation from the \code{\link[=create_dyadyears]{create_dyadyears()}} function
in this package.
}
\keyword{datasets}
