% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cow_igo_ndy}
\alias{cow_igo_ndy}
\title{Correlates of War Non-Directed Dyad-Year International Governmental Organizations (IGOs) Data}
\format{
A data frame with 917695 observations on the following 4 variables.
\describe{
\item{\code{ccode1}}{the Correlates of War state system code for the first state}
\item{\code{ccode2}}{the Correlates of War state system code for the second state}
\item{\code{year}}{the year}
\item{\code{dyadigos}}{the sum of mutual IGOs for which each state appears as a full member in a given year}
}
}
\usage{
cow_igo_ndy
}
\description{
This is a non-directed dyad-year version of the Correlates of War IGOs data. I use it internally for merging IGOs data into dyad-year data.
}
\details{
The \code{data-raw} directory on the project's Github contains additional information about how these data were generated from the otherwise
enormous dyad-year IGOs data provided by the Correlates of War project. Given the size of that data, and the size limitations of R packages for CRAN,
the data I provide here can only be simpler summaries. If you want specifics, you'll need to consult the underlying raw data provided on the Correlates
of War project.
}
\references{
Pevehouse, Jon C.W., Timothy Nordstrom, Roseanne W McManus, Anne Spencer Jamison, 2020. “Tracking Organizations in the World: The Correlates of War IGO Version 3.0 datasets”, Journal of Peace Research 57(3): 492-503.

Wallace, Michael, and J. David Singer. 1970. "International Governmental Organization in the Global System, 1815-1964." International Organization 24: 239-87.
}
\keyword{datasets}
