% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd-cow_majors.R
\docType{data}
\name{cow_majors}
\alias{cow_majors}
\title{Correlates of War Major Powers Data (1816-2016)}
\format{
A data frame with 14 observations on the following 8 variables.
\describe{
\item{\code{ccode}}{a numeric vector for the Correlates of War country code}
\item{\code{styear}}{the start year as a major power}
\item{\code{stmonth}}{the start montn as a major power}
\item{\code{stday}}{the start day as a major power}
\item{\code{endyear}}{the end year as a major power}
\item{\code{endmonth}}{the end month as a major power}
\item{\code{endday}}{the end day as a major power}
\item{\code{version}}{a version identifier}
}
}
\usage{
cow_majors
}
\description{
These are the Correlates of War major powers data.
}
\details{
Data are provided "as-is" with no additional re-cleaning before inclusion into this data set (beyond eliminating the state abbreviation).
}
\references{
Correlates of War Project. 2017. "State System Membership List, v2016." Online, \url{https://correlatesofwar.org/data-sets/state-system-membership}
}
\keyword{datasets}
