% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cow_states}
\alias{cow_states}
\title{Correlates of War State System Membership Data (1816-2016)}
\format{
A data frame with 243 observations on the following 10 variables.
\describe{
\item{\code{stateabb}}{a character vector for the state abbrevation}
\item{\code{ccode}}{a numeric vector for the Correlates of War country code}
\item{\code{statenme}}{a character vector for the state name}
\item{\code{styear}}{the start year in the system}
\item{\code{stmonth}}{the start month in the system}
\item{\code{stday}}{the start day in the system}
\item{\code{endyear}}{the end year in the system}
\item{\code{endmonth}}{the end month in the system}
\item{\code{endday}}{the end day in the system}
\item{\code{version}}{a version identifier}
}
}
\usage{
cow_states
}
\description{
These are the Correlates of War state system membership data.
}
\details{
Data are provided "as-is" with no additional re-cleaning before inclusion into this data set.
}
\references{
Correlates of War Project. 2017. "State System Membership List, v2016." Online, \url{https://correlatesofwar.org/data-sets/state-system-membership}
}
\keyword{datasets}
