% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cow.R
\docType{data}
\name{cow_trade_sy}
\alias{cow_trade_sy}
\title{Correlates of War National Trade Data Set (v. 4.0)}
\format{
A data frame with 14410 observations on the following four variables.
\describe{
\item{\code{ccode}}{the Correlates of War state system code}
\item{\code{year}}{the year}
\item{\code{imports}}{total imports of the state in current million USD}
\item{\code{exports}}{total exports of the state in current million USD}
}
}
\usage{
cow_trade_sy
}
\description{
These are state-year-level data for national trade from the Correlates of War
project.
}
\details{
The \code{data-raw} directory on the project's Github shows how the data were
processed.
}
\references{
Barbieri, Katherine and Omar M.G. Keshk. 2016. Correlates of War Project
Trade Data Set Codebook, Version 4.0. Online: \url{https://correlatesofwar.org}

Barbieri, Katherine, Omar M.G. Keshk, and Brian Pollins. 2009. "TRADING DATA:
Evaluating Our Assumptions and Coding Rules."
\emph{Conflict Management and Peace Science}, 26(5): 471-491.
}
\keyword{datasets}
