% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stateyears.R
\name{create_stateyears}
\alias{create_stateyears}
\title{Create state-years from state system membership data}
\usage{
create_stateyears(system = "cow", mry = TRUE, subset_years)
}
\arguments{
\item{system}{a character specifying whether the user wants Correlates of War
state-years ("cow") or Gleditsch-Ward ("gw") state-years. Correlates of War is the
default.}

\item{mry}{optional, defaults to TRUE. If TRUE, the function extends the
script beyond the most recent system membership updates to include observation
to the most recently concluded calendar year. For example, the Gleditsch-Ward
data extend to the end of 2017. When \code{mry == TRUE}, the function returns
more recent years (e.g. 2018, 2019) under the assumption that states alive at
the end of 2016 or 2020 are still alive today. Use with some care.}

\item{subset_years}{and optional character vector for subsetting the years
returned to just some temporal domain of interest to the user. For example,
\code{c(1816:1820)} would subset the data to just all state-years in 1816, 1817,
1818, 1819, and 1820. Be advised that it's easiest to subset the data after
the full universe of state-year data have been created. This means you could,
if you choose, effectively overwrite \code{mry = TRUE} with this argument since
the \code{mry} argument is applied at the expansion of the state system data
into state-year data.}
}
\value{
\code{create_stateyears()} takes state system membership data provided
by either Correlates of War or Gleditsch-Ward and returns a simple state-year
data frame. The Gleditsch-Ward state-years also include an indicator for
whether the state is a microstate.
}
\description{
\code{create_stateyears()} allows you to generate state-year data from either
the Correlates of War (CoW) state system membership data or the Gleditsch-Ward
(gw) system membership data.
}
\details{
The function leans on data made available in the \pkg{isard} package.
}
\examples{

# CoW is default, will include years beyond 2016 (most recent CoW update)
create_stateyears()

# Gleditsch-Ward, include most recent years
create_stateyears(system="gw")


}
\references{
Miller, Steven V. 2019. ``Create Country-Year and (Non)-Directed
Dyad-Year Data With Just a Few Lines in R''
\url{https://svmiller.com/blog/2019/01/create-country-year-dyad-year-from-country-data/}
}
\author{
Steven V. Miller
}
