% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-deprecated.R
\docType{data}
\name{data-deprecated}
\alias{data-deprecated}
\alias{cow_alliance}
\alias{ccode_democracy}
\alias{gwcode_democracy}
\title{Data sets that have been deprecated}
\format{
Users interested in the data referenced here can check the Github
repository associated with the package. The scripts that generated them are
available in the \verb{data-raw/} directory. Previous versions of the data are
available in CRAN archives as well.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 120784 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16731 rows and 5 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 18289 rows and 5 columns.
}
\usage{
cow_alliance

ccode_democracy

gwcode_democracy
}
\description{
These are data sets that have been deprecated and scheduled for
removal, or data that have since been removed after deprecation. Data sets
may be deprecated either by insistence of the data set's author, because they
will be relocated to another package for future development, or because the
data themselves are legacy data no longer in active demand or use in the
community. Deprecation and removal have the effect of also freeing up disk
space given CRAN's 5 MB limitation for R packages.
}
\details{
\code{cow_alliance} is defunct. The data set's maintainer requests
that users who want the Correlates of War alliance data in their analyses
should download and process the data manually, without assistance of any
convenience functions.

\code{ccode_democracy} is defunct. The data are now maintained in
the \pkg{isard} package as \code{cw_democracy}.

\code{gwcode_democracy} is defunct. The data are now maintained in
the \pkg{isard} package as \code{gw_democracy}.
}
\keyword{datasets}
\keyword{deprecated}
\keyword{internal}
