% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grh_arms_races}
\alias{grh_arms_races}
\title{Conventional Arms Races During Periods of Rivalry}
\format{
A data frame the following five variables.
\describe{
\item{\code{race_id}}{the arms race identifier}
\item{\code{ccode1}}{a numeric vector for the Correlates of War state code for the first state}
\item{\code{ccode2}}{a numeric vector for the Correlates of War state code for the second state}
\item{\code{styear}}{the start year for the arms race}
\item{\code{endyear}}{the end year for the arms race}
}
}
\usage{
grh_arms_races
}
\description{
This is a simple data set of 71 arms races reported by Gibler et al. in their
2005 article in \emph{Journal of Peace Research}.
}
\details{
Data are taken from the appendix of Gibler, Rider, and Hutchison's (2005)
article in \emph{Journal of Peace Research}. Read the article and appendix for
more information about coding procedures.
}
\references{
Gibler, Douglas M., Toby J. Rider, and Marc L. Hutchison. 2005. "Taking Arms Against
a Sea of Troubles: Conventional Arms Races during Periods of Rivalry"
\emph{Journal of Peace Research} 42(2): 131--47.
}
\keyword{datasets}
