% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gw_ddy}
\alias{gw_ddy}
\title{A directed dyad-year data frame of Gleditsch-Ward state system members}
\format{
A data frame with the following 3 variables.
\describe{
\item{\code{gwcode1}}{a numeric vector for the Correlates of War state code for the first state}
\item{\code{gwcode2}}{a numeric vector for the Correlates of War state code for the second state}
\item{\code{year}}{a numeric vector for the year}
}
}
\usage{
gw_ddy
}
\description{
This is a complete directed dyad-year data frame of Gleditsch-Ward
state system members. I offer it here as a shortcut for various other
functions. As a general rule, this data frame is updated after every
calendar year to include the most recently concluded calendar year.
}
\details{
Data are a quick generation from the \code{create_dyadyears(system="gw")} function in this package.
}
\keyword{datasets}
