% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hief}
\alias{hief}
\title{Historical Index of Ethnic Fractionalization data}
\format{
A data frame with 8808 observations on the following 5 variables.
\describe{
\item{\code{ccode}}{a Correlates of War state code}
\item{\code{gwcode}}{a Gleditsch-Ward state code}
\item{\code{year}}{the year}
\item{\code{efindex}}{a numeric vector for the estimate of ethnic fractionalization}
}
}
\usage{
hief
}
\description{
This is a data set with state-year estimates for ethnic fractionalization.
}
\details{
The \code{data-raw} directory on the project's Github contains more
information about how these data were created.
}
\references{
Drazanova, Lenka. 2020. "Introducing the Historical Index of Ethnic
Fractionalization (HIEF) Dataset: Accounting for Longitudinal Changes in
Ethnic Diversity." \emph{Journal of Open Humanities Data} 6:6
\doi{10.5334/johd.16}
}
\keyword{datasets}
