% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_capital_distance.R
\name{add_capital_distance}
\alias{add_capital_distance}
\title{Add capital-to-capital distance to a dyad-year data frame}
\usage{
add_capital_distance(data)
}
\arguments{
\item{data}{a dyad-year data frame (either "directed" or "non-directed")}
}
\value{
\code{add_capital_distance()} takes a dyad-year data frame and adds the
capital-to-capital distance between the first state and the second state. A minor
note about this function: cases of capital transition are recorded in the
\code{capitals} data but, in the conversion to capital-years (and eventual
merging into a dyad-year data frame), the Jan. 1 capital is used for calculating
distances.
}
\description{
\code{add_capital_distance()} allows you to add capital-to-capital
distance to a dyad-year data frame. The capitals are coded in the \code{capitals}
data frame, along with their latitudes and longitudes. The distance variable that
emerges \code{capdist} is calculated using the "Vincenty" method (i.e. "as the crow
flies") and is expressed in kilometers.
}
\examples{

library(magrittr)
cow_ddy \%>\% add_capital_distance()


}
\author{
Steven V. Miller
}
