% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_strategic_rivalries.R
\name{add_strategic_rivalries}
\alias{add_strategic_rivalries}
\title{Add Thompson and Dreyer's (2012) strategic rivalry data to dyad-year data frame}
\usage{
add_strategic_rivalries(data)
}
\arguments{
\item{data}{a dyad-year data frame (either "directed" or "non-directed")}
}
\value{
\code{add_strategic_rivalries()} takes a dyad-year data frame and adds information about ongoing strategic rivalries.
}
\description{
\code{add_strategic_rivalries()} merges in Thompson and Dreyer's (2012) strategic rivalry data
to a dyad-year data frame. The right-bound, as of right now, are bound at 2010.
}
\details{
\code{add_strategic_rivalries()} will include some other information derived from the rivalry data that the
user may not want (e.g. start year of the rivalry). Feel free to select those out after the fact. Function includes
an on-the-fly adjustment for Austria for rivalry #79. In this case, the Austria-Serbia rivalry continues for two years
after Austria-Hungary (\code{ccode: 300}) became Austria (\code{ccode: 305}).
}
\examples{

# just call `library(tidyverse)` at the top of the your script
library(magrittr)
cow_ddy \%>\% add_stragegic_rivalries()

}
\references{
Miller, Steven V. 2019. "Create and Extend Strategic (International) Rivalry Data in R".
URL: \url{http://svmiller.com/blog/2019/10/create-extend-strategic-rivalry-data-r/}

Thompson, William R. and David Dreyer. 2012. \emph{Handbook of International Rivalries}. CQ Press.
}
\author{
Steven V. Miller
}
