% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ucdp_acd.R
\name{add_ucdp_acd}
\alias{add_ucdp_acd}
\title{Add UCDP Armed Conflict Data to state-year data frame}
\usage{
add_ucdp_acd(data, type, issue)
}
\arguments{
\item{data}{state-year data frame}

\item{type}{the types of armed conflicts the user wants to consider, specified as a character vector.
Options include "extrasystemic", "interstate", "intrastate", and "II". "II" is convenience shorthand for
"internationalized intrastate". If you want just one (say: "intrastate"), then the type you want in quotes
is sufficient. If you want multiple, wrap it in a vector with \code{c()}.}

\item{issue}{do you want to subset the data to just different armed conflicts over different types of issues?
If so, specify those here as you would with the \code{type} argument. Options include "territory", "government",
and "both".}
}
\value{
\code{add_ucdp_acd()} takes a state-year data frame and returns state-year information from the
UCDP Armed Conflict data set (v. 20.1). The variables returned are whether there is an ongoing armed
conflict in that year, whether there was an armed conflict episode onset that year, what was the maximum
intensity observed that year (if an armed conflict was observed), and a character vector of the associated
conflict IDs that year.
}
\description{
\code{add_ucdp_acd()} allows you to add UCDP Armed Conflict data to a state-year data frame
}
\details{
Right now, only state-year data are supported. Function is in true pilot mode.
}
\references{
Gleditsch, Nils Petter; Peter Wallensteen, Mikael Eriksson, Margareta Sollenberg & Håvard Strand (2002)
Armed Conflict 1946–2001: A New Dataset. \emph{Journal of Peace Research} 39(5): 615–637.

Pettersson, Therese; Stina Högbladh & Magnus Öberg (2019). Organized violence, 1989-2018 and peace
agreements. Journal of Peace Research 56(4): 589-603.
}
\author{
Steven V. Miller
}
