% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_leaderyears.R
\name{create_leaderyears}
\alias{create_leaderyears}
\title{Create leader-years from leader data}
\usage{
create_leaderyears(system = "archigos", standardize_cow = TRUE)
}
\arguments{
\item{system}{a leader system with which to create leader-years. Right now, only "archigos" is supported.}

\item{standardize_cow}{logical, defaults to TRUE. If TRUE, the function standardizes the leader years to just those that
overlap  with state system membership in the Correlates of War state system (see: \code{cow_states}). If FALSE, the function
returns all leader-years as implied by the Archigos data.}
}
\value{
\code{create_leaderyears()} takes leader-level data available in \pkg{peacesciencer}
and returns a leader-year-level data frame. This minimal output contains the observation ID
from Archigos, the year, the Correlates of War state code for the leader (i.e. leaders are nested in states),
the leader's name in Archigos (if it may help the reader to have that), an approximation of the leader's age,
and the year in office for the leader (as a running count, starting at 1).
}
\description{
\code{create_leaderyears()} allows you to generate leader-year data from
leader-level data provided in \pkg{peacesciencer}
}
\details{
\code{create_leaderyears()}, as of writing, only supports the Archigos data set of leaders. Importantly:
the absence of much leader-level covariates (of which I am aware) means, for now, the data that are returned are
treated as observationally equivalent to state-year data. Users should be careful here, but it does mean the data
will work with other functions in \pkg{peacesciencer} that have support for state-year data (e.g. \code{add_nmc()},
\code{add_rugged_terrain()}). This is declared in the attribute field.

Many leader ages are known with precision. Many are not recorded in the Archigos data. Knowing well that years are aggregates
of days, the leader age variable that gets returned in this output should be treated as an approximation of the leader's age.
}
\examples{
\donttest{
create_leaderyears()

create_leaderyears(standardize_cow = FALSE)
}

}
\references{
Goemans, Henk E., Kristian Skrede Gleditsch, and Giacomo Chiozza. 2009. "Introducing Archigos: A Dataset of Political Leaders"
\emph{Journal of Peace Research} 46(2): 269--83.
}
\author{
Steven V. Miller
}
