% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-deprecated.R
\docType{data}
\name{data-deprecated}
\alias{data-deprecated}
\alias{cow_alliance}
\alias{gml_dirdisp}
\alias{gml_mid_ddydisps}
\alias{gml_mid_disps}
\alias{gml_part}
\alias{gml_mid_ddlydisps}
\alias{gml_mid_dirleaderdisps}
\title{Data sets that have been deprecated}
\format{
Users interested in the data referenced here can check the Github
repository associated with the package. The scripts that generated them are
available in the \verb{data-raw/} directory. Previous versions of the data are
available in CRAN archives as well.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 120784 rows and 7 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10276 rows and 39 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9284 rows and 24 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2174 rows and 11 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5217 rows and 19 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10708 rows and 16 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11686 rows and 16 columns.
}
\usage{
cow_alliance

gml_dirdisp

gml_mid_ddydisps

gml_mid_disps

gml_part

gml_mid_ddlydisps

gml_mid_dirleaderdisps
}
\description{
These are data sets that have been deprecated and scheduled for
removal, or data that have since been removed after deprecation. Data sets
may be deprecated either by insistence of the data set's author, because they
will be relocated to another package for future development, or because the
data themselves are legacy data no longer in active demand or use in the
community. Deprecation and removal have the effect of also freeing up disk
space given CRAN's 5 MB limitation for R packages.
}
\details{
\code{cow_alliance} is defunct. The data set's maintainer requests
that users who want the Correlates of War alliance data in their analyses
should download and process the data manually, without assistance of any
convenience functions.

\code{gml_dirdisp} is defunct. One of the data set's maintainers
requests that users who want to use the data should download and process the
data manually, without assistance of any convenience functions or software
that ultimately suggest design decisions on behalf of the researcher.

\code{gml_mid_ddydisps} is defunct. One of the data set's maintainers
requests that users who want to use the data should download and process the
data manually, without assistance of any convenience functions or software
that ultimately suggest design decisions on behalf of the researcher.

\code{gml_mid_disps} is defunct. One of the data set's maintainers
requests that users who want to use the data should download and process the
data manually, without assistance of any convenience functions or software
that ultimately suggest design decisions on behalf of the researcher.

\code{gml_part} is defunct. One of the data set's maintainers
requests that users who want to use the data should download and process the
data manually, without assistance of any convenience functions or software
that ultimately suggest design decisions on behalf of the researcher.

\code{gml_mid_ddlydisps} is defunct. One of the data set's maintainers
requests that users who want to use the data should download and process the
data manually, without assistance of any convenience functions or software
that ultimately suggest design decisions on behalf of the researcher.

\code{gml_mid_dirleaderdisps} is defunct. One of the data set's maintainers
requests that users who want to use the data should download and process the
data manually, without assistance of any convenience functions or software
that ultimately suggest design decisions on behalf of the researcher.
}
\keyword{datasets}
\keyword{deprecated}
\keyword{internal}
