% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_extdata.R
\name{download_extdata}
\alias{download_extdata}
\title{Download Some Extra Data for Peace Science Research}
\usage{
download_extdata(overwrite = FALSE)
}
\arguments{
\item{overwrite}{logical, defaults to FALSE. If FALSE, the function checks to see if you've already
downloaded the data and, if you already have, it does nothing. If TRUE, the function redownloads the data.}
}
\value{
\code{download_extdata()} downloads some extra data stored on my website (\url{http://svmiller.com}) and
sticks them in the \code{extdata} directory in the package. Right now, these data are just the directed dyad-year
Correlates of War trade data.
}
\description{
\code{download_extdata()} leverages R's \code{inst} directory flexibility to allow you to download
some extra data and store it in the package.
}
\section{A Description of Various Data Sets This Will Download}{
Running \code{download_extdata()} returns the following data that will be stored in the package's \code{extdata} directory.
\subsection{Correlates of War Dyadic Trade Data Set (v. 4.0)}{

These are directed dyad-year-level data for dyadic trade from the Correlates of War project. The trade values presented here
have been rounded to three decimal points to conserve space. The data downloaded by this function are about 4.1 megabytes in size.\tabular{ll}{
   COLUMN \tab DESCRIPTION \cr
   \code{ccode1} \tab a numeric vector for the Correlates of War state code for the first state \cr
   \code{ccode2} \tab a numeric vector for the Correlates of War state code for the second state \cr
   \code{year} \tab the year \cr
   \code{flow1} \tab imports of \code{ccode1} from \code{ccode2}, in current million USD \cr
   \code{flow2} \tab imports of \code{ccode2} from \code{ccode1}, in current million USD \cr
   \code{smoothflow1} \tab smoothed \code{flow1} values \cr
   \code{smoothflow2} \tab smoothed \code{flow2} values \cr
}

}

\subsection{Directed Leader Dyad-Year Data, 1870-2015 (CoW States)}{

These are all directed leader dyad-year data from 1870-2015. Data come from the Archigos data (version 4.1). The
data are standardized to just those observations where both leaders and states appear in the CoW state system data. The
data downloaded by this function are about 2 megabytes in size.\tabular{ll}{
   COLUMN \tab DESCRIPTION \cr
   \code{year} \tab the year \cr
   \code{obsid1} \tab the unique Archigos (v. 4.1) observation ID for the first leader \cr
   \code{obsid2} \tab the unique Archigos (v. 4.1) observation ID for the second leader \cr
   \code{ccode1} \tab a numeric vector for the Correlates of War state code for the first state \cr
   \code{ccode2} \tab a numeric vector for the Correlates of War state code for the second state \cr
   \code{gender1} \tab the gender of \code{obsid1} ("M" or "F") \cr
   \code{gender2} \tab the gender of \code{obsid2} ("M" or "F") \cr
   \code{leaderage1} \tab the approximate age (i.e. \code{year - yrborn}) for \code{obsid1} in the year \cr
   \code{leaderage2} \tab the approximate age (i.e. \code{year - yrborn}) for \code{obsid2} in the year \cr
   \code{yrinoffice1} \tab a running count for the tenure of \code{obsid1}, starting at 1. \cr
   \code{yrinoffice2} \tab a running count for the tenure of \code{obsid2}, starting at 1. \cr
}

}

\subsection{Directed Leader Dyad-Year Data, 1870-2015 (Gleditsch-Ward States)}{

These are all directed leader dyad-year data from 1870-2015. Data come from the Archigos data (version 4.1). The
data represent every possible dyadic leader-pairing in the Archigos data (which is denominated in the Gleditsch-Ward
system). The data downloaded by this function are about 2.2 megabytes in size.\tabular{ll}{
   COLUMN \tab DESCRIPTION \cr
   \code{year} \tab the year \cr
   \code{obsid1} \tab the unique Archigos (v. 4.1) observation ID for the first leader \cr
   \code{obsid2} \tab the unique Archigos (v. 4.1) observation ID for the second leader \cr
   \code{gwcode1} \tab a numeric vector for the Gleditsch-Ward state code for the first state \cr
   \code{gwcode2} \tab a numeric vector for the Gleditsch-Ward state code for the second state \cr
   \code{gender1} \tab the gender of \code{obsid1} ("M" or "F") \cr
   \code{gender2} \tab the gender of \code{obsid2} ("M" or "F") \cr
   \code{leaderage1} \tab the approximate age (i.e. \code{year - yrborn}) for \code{obsid1} in the year \cr
   \code{leaderage2} \tab the approximate age (i.e. \code{year - yrborn}) for \code{obsid2} in the year \cr
   \code{yrinoffice1} \tab a running count for the tenure of \code{obsid1}, starting at 1. \cr
   \code{yrinoffice2} \tab a running count for the tenure of \code{obsid2}, starting at 1. \cr
}

}
}

\examples{

\dontrun{
# Here's where the data are going to be downloaded.
system.file("extdata", package="peacesciencer")
# Now, let's download the data.
download_extdata()
}
}
\references{
Barbieri, Katherine, Omar M. G. Keshk, and Brian Pollins. 2009. "TRADING DATA: Evaluating our Assumptions and Coding Rules."
\emph{Conflict Management and Peace Science}. 26(5): 471-491.

Goemans, Henk E., Kristian Skrede Gleditsch, and Giacomo Chiozza. 2009. "Introducing Archigos: A Dataset of Political Leaders"
\emph{Journal of Peace Research} 46(2): 269--83.
}
\author{
Steven V. Miller
}
