% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_extdata.R
\name{download_extdata}
\alias{download_extdata}
\title{Download Some Extra Data for Peace Science Research}
\usage{
download_extdata(overwrite = FALSE)
}
\arguments{
\item{overwrite}{logical, defaults to FALSE. If FALSE, the function checks to see if you've already
downloaded the data and, if you already have, it does nothing. If TRUE, the function redownloads the data.}
}
\value{
\code{download_extdata()} downloads some extra data stored on my website (\url{http://svmiller.com}) and
sticks them in the \code{extdata} directory in the package. Right now, these data are just the directed dyad-year
Correlates of War trade data.
}
\description{
\code{download_extdata()} leverages R's \code{inst} directory flexibility to allow you to download
some extra data and store it in the package.
}
\section{A Description of Various Data Sets This Will Download}{
Running \code{download_extdata()} returns the following data that will be stored in the package's \code{extdata} directory.
\subsection{Correlates of War Dyadic Trade Data Set (v. 4.0)}{

These are directed dyad-year-level data for national trade from the Correlates of War project.\tabular{ll}{
   COLUMN \tab DESCRIPTION \cr
   \code{ccode1} \tab a numeric vector for the Correlates of War state code for the first state \cr
   \code{ccode2} \tab a numeric vector for the Correlates of War state code for the second state \cr
   \code{year} \tab the year \cr
   \code{flow1} \tab imports of \code{ccode1} from \code{ccode2}, in current million USD \cr
   \code{flow2} \tab imports of \code{ccode2} from \code{ccode1}, in current million USD \cr
   \code{smoothflow1} \tab smoothed \code{flow1} values \cr
   \code{smoothflow2} \tab smoothed \code{flow2} values \cr
}

}

\subsection{Directed Leader Dyad-Year Data, 1870-2015}{

These are all directed leader dyad-year data from 1870-2015. Data come from the Archigos data (version 4.1).\tabular{ll}{
   COLUMN \tab DESCRIPTION \cr
   \code{year} \tab the year \cr
   \code{leadid1} \tab the unique Archigos (v. 4.1) leader ID for the first leader \cr
   \code{leadid2} \tab the unique Archigos (v. 4.1) leader ID for the second leader \cr
   \code{ccode1} \tab a numeric vector for the Correlates of War state code for the first state \cr
   \code{startdate1} \tab the start date in office for the first leader \cr
   \code{enddate1} \tab the end date in office for the first leader \cr
   \code{ccode2} \tab a numeric vector for the Correlates of War state code for the second state \cr
   \code{startdate2} \tab the start date in office for the second leader \cr
   \code{enddate2} \tab the end date in office for the second leader \cr
}

}
}

\examples{

\dontrun{
# Here's where the data are going to be downloaded.
system.file("extdata", package="peacesciencer")
# Now, let's download the data.
download_extdata()
}
}
\references{
Barbieri, Katherine, Omar M. G. Keshk, and Brian Pollins. 2009. "TRADING DATA: Evaluating our Assumptions and Coding Rules."
\emph{Conflict Management and Peace Science}. 26(5): 471-491.

Goemans, Henk E., Kristian Skrede Gleditsch, and Giacomo Chiozza. 2009. "Introducing Archigos: A Dataset of Political Leaders"
\emph{Journal of Peace Research} 46(2): 269--83.
}
\author{
Steven V. Miller
}
