% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gw_states}
\alias{gw_states}
\title{Gleditsch-Ward (Independent States) System Membership Data (1816-2017)}
\format{
A data frame with 216 observations on the following 5 variables.
\describe{
\item{\code{gwcode}}{a numeric vector for the Gleditsch-Ward country code}
\item{\code{stateabb}}{a character vector for state abbreviation}
\item{\code{statename}}{a character vector for the state name}
\item{\code{startdate}}{the start date in the data}
\item{\code{enddate}}{the end date in the data}
}
}
\usage{
gw_states
}
\description{
These are the independent states in Gleditsch and Ward's data.
}
\details{
Data originally provided by Gleditsch with no column names. Column names
were added before some light re-cleaning in order to generate these data.
"Wuerttemberg" and "Cote D'Ivoire" in the \code{statename} column needed to be
renamed to ensure maximal compliance with CRAN, which raises notes for
every non-ASCII character that appears in its package. I do not think this
to be problematic at all and, after all, state names should never be
a basis for something like a match or merge you would do in
\pkg{countrycode}.
}
\references{
Gleditsch, Kristian S. and Michael D. Ward. 1999. "A Revised List of
Independent States since the Congress of Vienna."
\emph{International Interactions} 25(4): 393--413.
}
\keyword{datasets}
